+++ phase_two/models.py
@@ -32,6 +32,6 @@ class Group(BaseGroup):
 
 
 class Player(BasePlayer):
-    cutoff = models.DecimalField(blank=False, max_digits=4, decimal_places=2)
+    cutoff = models.FloatField(blank=False)
     auction = models.IntegerField(blank=False)
     clicked = models.IntegerField()
